requirements.txt
llama-index==0.9.48
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1